+++ ios/Classes/SwiftFlutterHtmlToPdfPlugin.swift
@@ -21,6 +21,7 @@ public class SwiftFlutterHtmlToPdfPlugin: NSObject, FlutterPlugin{
         let height = Double(args!["height"] as! Int)
         let orientation = args!["orientation"]
         let margins = args!["margins"] as? [Int]
+        let printSize = args!["printSize"] as? String
         
         let viewControler = UIApplication.shared.delegate?.window?!.rootViewController
         wkWebView = WKWebView.init(frame: CGRect(origin: CGPoint(x:0, y:0), size: CGSize(width:width, height: height)))