+++ lib/flutter_html_to_pdf_plus.dart
@@ -1,5 +1,3 @@
-library flutter_html_to_pdf_plus;
-
 export 'src/flutter_html_to_pdf_plus.dart';
 export 'src/pdf_print_configuration.dart';
 export 'src/print_config_enums.dart';