+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+# 0.4.0
+
+### New features
+
+- Add RTL (Right-to-Left) text direction support
+
+# 0.3.0
+
+### New features
+
+- Support for desktop (macOS)
+
+### Bug fixes
+
+- Migrate Android to use Gradle 8 & kts
+
+# 0.2.0
+
+### New features
+
+- Support for custom document size
+
+### Bug fixes
+
+- Fix margin configuration not taken into account on iOS
+
+### Misc
+
+- Update README
+
 # 0.1.0
 
 Initial release of this updated version of the plugin (forked from [flutter_html_to_pdf](https://github.com/afur/flutter_html_to_pdf))